+++ kodilla-patterns2/src/main/java/com/kodilla/patterns2/adapter/bookclassifier/libraryb/Statistics.java
@@ -5,9 +5,9 @@ import java.util.Map;
 
 public class Statistics implements BookStatistics {
     @Override
-    public int averagePublicationYear(Map<BookSignature, BookB> books) {
+    public double averagePublicationYear(Map<BookSignature, BookB> books) {
         if (books.size() == 0) return 0;
-        int sum = 0;
+        double sum = 0;
         for (Map.Entry<BookSignature, BookB> entry : books.entrySet()) {
             sum += entry.getValue().getYearOfPublication();
         }
@@ -15,9 +15,9 @@ public class Statistics implements BookStatistics {
     }
 
     @Override
-    public int medianPublicationYear(Map<BookSignature, BookB> books) {
+    public double medianPublicationYear(Map<BookSignature, BookB> books) {
         if (books.size() == 0) return 0;
-        int[] years = new int[books.size()];
+        double[] years = new double[books.size()];
         int n = 0;
         for (Map.Entry<BookSignature, BookB> entry : books.entrySet()) {
             years[n] = entry.getValue().getYearOfPublication();
